[BigQuery] Add BigQueryQueryProvider interface (Recidiviz/recidiviz-data#29101)
## Description of the change

Introduces a new simple `BigQueryQueryProvider` interface that can be
used to access a BQ query, as well as a few basic implementations.

There's no real reason why the queries we run in pipelines need to be
derived from a deployed BQ view builder, so this interface will be used
by the data loading code so that we have the option to swap out for, for
example, a `BigQueryQueryBuilder` later (see Recidiviz/recidiviz-data#29103).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#22528

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 491f0ea09a3d73ae2aea85cc21fa5609e037bc23