[Dataflow] Some minor testing cleanup (Recidiviz/recidiviz-data#29102)
## Description of the change

Makes some minor changes to how mocking is set up in pipeline tests in
advance of reading reference view queries directly in tests:
1) Mocks out project_id in all pipeline tests - this will be important
because we will need project_id set to build view builders inside of
pipelines.
2) Mocks `apache_beam.io.ReadFromBigQuery` directly instead of mocking
the `ReadFromBigQuery` class. In a subsequent PR the file that our
ReadFromBigQuery call happens in will change again so this simplifies
the test changes when that happens.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#22528

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: edea90844a83027e93de7627199b99a7c9a1b606