[US_AZ] Do not infer or overwrite misaligned dates (Recidiviz/recidiviz-data#29171)
## Description of the change
In the task deadlines query, I am currently overwriting dates from the
raw data in two instances:
- when an eligibility date is after a due date, because this is not
allowed by platform validations, I am making the two dates equal based
on the assumption that the person was eligible when they were released.
- when there is no eligibility date, but a person has already been
released on TRP, I am currently overwriting the NULL eligibility date
with the actual release date based on the assumption that the person was
eligible when they were released.

Shalin and Annie requested:
- Leave eligibility dates after due dates whenever that reverse-ordering
appears in the raw data
- Leave eligibility dates null when they are null, even when a person
has been released

The first of these requests is not possible, because of the validations
that the entity must pass in order for ingest to run. To get around
this, I will fill in the eligibility date with "1000-01-01" whenever it
is actually after the due date.

Summary of possible values after this change:
- If a person has a NULL eligible_date and a NULL due_date, that means
they are approved or tentatively approved for TPR, but do not have a
planned release date, and have not actually been released.
- If a person has a NULL eligible_date and a valued due_date, that means
they were released on TPR but never had a planned TPR release date.
- If a person has an eligible_date of "1000-01-01" and a valued
due_date, that means they were released on TPR on the due date, but
their planned/earliest possible TPR release date was after that date. We
see this when a person has been released on transition release more than
once on the same incarceration stint; it's unclear if that is a data
entry error or a legitimate possibility. The releases are typically >90
days apart, so I'm inclined to think its an error in those cases. This
reverse-dating could be a legitimate feature of a person's release in
some other cases?
- If a person has a valued eligible_date and a NULL due_date, it means
they have a planned TPR release date, but have not been released on TPR.
This could be because they are still incarcerated, or because they were
ultimately released _not_ on TPR despite being approved for it.

Tests:

- [x] view query is deterministic
- [x] each pair of person_id and update_datetime has exactly one row
- [x] run_mappings script passes
- [x] sandbox ingest pipeline
[succeeds](https://console.cloud.google.com/dataflow/jobs/us-west1/2024-04-19_14_05_44-1797896327962440688;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

I also removed some unnecessary datetime SQL parsers in the
`AZ_DOC_SC_OFFENSE` raw data config file here to make it easier to test
the ingest view query. That required that I update a fixture file for
`state_sentence_group`, which also uses that raw data table for ingest
view tests to pass. The results were not meaningfully different.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 57b46e496d2a03582a457e8fd2d188b761a06e09