[raw data import dag] normalization pass (Recidiviz/recidiviz-data#29024)
## Description of the change

normalization pass implementation for the raw data import dag (apologies
for the size of the PR). a quick normalization pass refresher: this step
is responsible for making any csv file a state will give us into a
big-query readable csv. it does so by treating the file as a text stream
instead of csv (we dont care about teh contents right now, just the
form).

a few things to note about this PR:
- creates `BytesChunkReader` which just wraps around the `.read` method
to only read from the byte start/end provided
- BigQuery's
[docs](https://cloud.google.com/bigquery/docs/loading-data-cloud-storage-csv#field_delimiters)
say the following about field delimiters
> Delimiters in CSV files can be any single-byte character. If the
source file uses ISO-8859-1 encoding, any character can be a delimiter.
If the source file uses UTF-8 encoding, any character in the decimal
range 1-127 (U+0001-U+007F) can be used without modification. You can
insert an ISO-8859-1 character outside of this range as a delimiter, and
BigQuery will interpret it correctly. However, if you use a multibyte
character as a delimiter, some of the bytes will be interpreted
incorrectly as part of the field value.
> Generally, it's a best practice to use a standard delimiter, such as a
tab, pipe, or comma. The default is a comma.

  i interpreted this as basically saying:
>"for fixed length encodings we accept (i.e. all but utf-8), any
delimiter works but for variable-length encodings (i.e. utf), the
delimiter must be a single byte (i.e. code points 1-127)".

i use fixed vs variable length here because the phrase "multibyte
character as a delimiter" would mean that _all_ characters would be
invalid for a encoding set that is actually multiple bytes like UTF-16
which isn't true; instead i think they mean multibyte either for
variable-length encodings or just for utf-8 (which are the same rn)

as a result, for `is_big_query_valid_delimiter`, i had considered using
`ord(delimiter) <= 127` instead of `len(...) == 1` since it more closely
aligns w/ the docs, but opted for the second as it is easier to
understand that `ord`
- i tried to abstract away encoding aliases (i.e. `latin-1` vs
`iso8859-1`) using `codecs.lookup` to make things standardized
- `headers` -- since we are reading in chunks of a file that don't have
headers AND big query assumes that all csv files either have headers or
not, we either need to decide to have all files have no headers or add
headers to the chunks that are not the 0-th chunk or have
`infer_columns_from_config==True`. in this PR, i opted for all chunk
outputs to have headers via the explicit passing of `headers` as a
string on the `RequiresNormalizationFileChunk` object that will get
validated during the `AccountingPass` step. i could also imagine us
generating the headers instead of passing them in like we do for
`infer_columns_from_config=False` (i.e. `columns_from_file_config =
[column.name for column in file_config.columns]`) since we would have
validated them at this point but opted for `headers` right now to make
it clearer why / what we are doing what we are doing at this stage

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#28586
Closes Recidiviz/recidiviz-data#29152

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3d507e915cd6b42fd429d5647c209a0099a49d29