[Justice Counts] don't default enabled to false for supervision metrics when there are no subsystems (Recidiviz/recidiviz-data#29288)
## Description of the change

During playtesting for the new admin panel tool to copy over metric
config from a superagency to it's child agencies, the following bug came
up:

```
Agency: Community Solutions, Inc (I'm still lol'ing at this name)
Metric: Funding (in Supervision sector)
Metric is marked as unavailable in superagency, but after copy is still available in all child agencies.
```

Turns out that this agency is a Supervision agency with no subsystems.
There is metric config logic where we infer that if that
`disaggregated_by_supervision_subsystems` is False (this is a fine
assumption FWIW) if it has not been explicitly set. Later on, on line
932 of datapoint.py we set `enabled` to the metric to ` not
disaggregated_by_supervision_subsystems`. This is fine, but since agency
doesn't have any supervision subsystems we would overwrite the disabled
metric to be enabled. I fixed this bug by just checking if the agency
had supervision subsystems. If it does not, we don't execute the logic
on line 932.

I also went ahead and brought back the
`recidiviz/justice_counts/jobs/copy_superagency_metric_settings_to_child_agencies.py`
that allows me to copy over metrics locally. This was an invaluable tool
for me as I did this debugging. I also refactored the job so that there
was no repeated code between the scripts.

This only solves the first part of the ticket, the other error described
is a separate issue and I will solve that in another PR.

I tested this locally and confirmed that I was successfully able to copy
over the metrics from the superagency to its child agencies.

## Related issues
Recidiviz/recidiviz-data#28871

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 594362493e500ef25eaca847dcbb8c77582f5bbd