[US_OR] Switching some supervision and incarceration periods (Recidiviz/recidiviz-data#28530)
## Description of the change

Originally we used location type to determine if a period is a
supervision or incarceration period because they use the same tables and
the CURRENT_STATUS does not have the same level of information/accuracy.
However, some periods in Oregon have a CURRENT_STATUS of IN (adult in
custody/facility) or LC (local jail) with a location that is not a
prison or jail (usually a community location). OR says these should
still be counted as Incarceration Periods if the CURRENT_STATUS is IN or
LC, so this change removed those periods from supervision periods and
added them to incarceration periods.

The supervision period results went from 2448054 to 2447002 and the
incarceration period results went from 2356469 to 2368316.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c908e9ba8d75e38f4d771afc6d0008b81981f04c