[Justice Counts] Replace INVALID_CHILD_AGENCY error with UNEXPECTED_ERROR (Recidiviz/recidiviz-data#29387)
## Description of the change

Replace INVALID_CHILD_AGENCY error with UNEXPECTED_ERROR.

The codebase is configured right now to send a INVALID_CHILD_AGENCY
error when we experience an error that we don't know how to handle.

This INVALID_CHILD_AGENCY is misleading, and it seems like it's an
accident where it is. Let's replace it with a more straightforward
UNEXPECTED_ERROR error. This is an improvement because it tells us more
directly that we are not handling the error properly - instead of
leading us down the wrong path by suggesting something is up with our
child agency logic.

## Testing
I tested this locally and got the unexpected error message below.

<img width="967" alt="Screenshot 2024-04-26 at 10 17 28 AM"
src="https://github.com/Recidiviz/recidiviz-data/assets/130382407/f428a24b-fb79-48b3-8aaf-af7235aadda7">

Zenhub Ticket Recidiviz/recidiviz-data#29303

GitOrigin-RevId: 10ac0975afe116189831f9129b4ef98a4f775703